+++ header/Hamming.h
@@ -19,7 +19,7 @@ class Hamming : public ProcessData {
     public:
         Hamming(std::queue<uint8_t>* buffInput
             , std::queue<uint8_t>* buffOutput
-            , int bufferSize);
+            , size_t bufferSize);
     private:
         void iteration() override;
 };
