+++ header/Interleaving.h
@@ -13,7 +13,7 @@ class Interleaving : public ProcessData {
     public:
         Interleaving(std::queue<uint8_t>* buffInput
             , std::queue<uint8_t>* buffOutput
-            , int bufferSize);
+            , size_t bufferSize);
     private:
         void iteration() override;
 };
